Started new difficulty system
IMPORTANT: I need to move this to the enemies script, as thats on a "solid" gameobject. When it's all being run in the shoot script, the values reset and are different for each bullet. It's 9pm and I was dumb and made mistakes. Please forgive, future me, I know you're tired. Get some food, then solve this problem. <3